Added a pair of unit tests for roundtripping various values.
'numbers' creates a table with a column of each integer and
floating-point type, inserts various values (using the text and
the binary protocols), and reads them out (again using the text
and binary protocols). It should probably also test DECIMAL and
NUMERIC and some of the oddballs like YEAR.

'nulls' fills a table with null and non-null values in a specific
pattern and then reads them out.