OpenSSL: Fix broken dependency configure.ac script (add ssl feature)
Previously, the code was checking that OpenSSL was present, and aborting compilation otherwise, but not actually doing anything.
The code expected the macro `USE_OPENSSL` to be defined, which wasn't, and the binary wasn't linked with `-lssl` either.

The configure script already provide a `--with-openssl=<path>` option, but it doesn't allow disabling the feature.
Changing it to be `--with-openssl=<no>|<path>[|yes]` would make it quite weird to use, so instead we separate enabling the *feature* of using OpenSSL to the actual dependency.
This could also be extended to use another crypto library in the future.