Expand and fix string escaping capabilities
The mysql manual lists 11 characters which should be escaped,
including 2 which should only escaped for string using in pattern-matching context (e.g. LIKE clause).
However, `drizzle_escape_string` was only escaping 7 characters.
As changing `drizzle_escape_string` is not an option, a new `drizzle_escape_str` is introduced
which takes an extra parameter, for specifying whether the string will be
used in a pattern-matching context.