Correctly set the SONAME for libdrizzle-redux6
The previous attempt to do so (#150) only modified the package name and binary name after the fact.
However, when generating shared library, a special ELF field (DT_SONAME) is written, which includes the binary's name.
When linking, this field (DT_SONAME) is used to populate the ELF field DT_NEEDED of the binary,
which is later used by the dynamic loader (ld.so) to load binaries on demand.
As the previous attempt was moving the file, DT_NEEDED was still populated with libdrizzle-redux,
but ld.so couldn't find it in the file system, so the binary wouldn't run.

For further reference, https://autotools.io/libtool/version.html can be used as a reference (see '4.3. Multiple libraries versions').

Note that the major version number is now applied to AC_INIT directly instead of after-the-fact,
as `AC_INIT` requires literals to be used, as it can be expanded multiple times, early on.

The deb and rpm fixes were authored by @andreas-bok-sociomantic.
The installation path for dev headers was updated to `/usr/include/libdrizzle-redux[MAJOR]/libdrizzle-redux`.