[TRAVIS] Move docker image builds to before_script step
Separate installation of depencies on the native travis
build environment and operations related to the build
iteself. `before_install` should once concern the former and
`before_script` the latter.

It allows us to easily skip steps when defining the build stages,
e.g. not building the docker images if it isn't used in the
stage